TRayIcon && Salidas - hasta

+++ Libraries/SIM900/SIM900.cpp
@@ -94,13 +94,13 @@ void SIM900::PowerOn()
 
 /////// ENDMETHODS
 
- bool SIM900::GetHttpBuffer(uint8_t desde,uint8_t length)
+ bool SIM900::GetHttpBuffer(uint16_t desde,uint16_t length)
  {
 	return (SendCommandCheck( F("AT+HTTPREAD=%i,%i"),F("+HTTPREAD:"),desde,length)==RX_CHECK_OK);
  }
  
  
- bool SIM900::GetHttpBuffer(uint8_t desde,uint8_t length,char *buffer)
+ bool SIM900::GetHttpBuffer(uint16_t desde,uint16_t length,char *buffer)
  {
 	 if (GetHttpBuffer(desde,length))
 	 {
@@ -115,7 +115,7 @@ void SIM900::PowerOn()
 		
  }
  
- bool SIM900::GetHttpBuffer(uint8_t desde,uint8_t length,char *buffer,const __FlashStringHelper *compare)
+ bool SIM900::GetHttpBuffer(uint16_t desde,uint16_t length,char *buffer,const __FlashStringHelper *compare)
  {
 	 
 	 if (GetHttpBuffer(desde,length,buffer))
